refactor(#251): entities 단계에 public API 추가
