+++ src/widgets/home/FavoriteGoalsSection/FavoriteGoalsSection.tsx
@@ -3,7 +3,7 @@
 import { useRouter } from "next/navigation";
 import { useEffect, useRef, useState } from "react";
 
-import { FavoriteGoalItem } from "@/entities/goal/types/favorite.types";
+import { FavoriteGoalItem } from "@/entities/goal";
 import { useInfiniteScroll } from "@/shared/hooks/useInfiniteScroll/useInfiniteScroll";
 import Button from "@/shared/ui/Button/Button/Button";
 import { Icon } from "@/shared/ui/Icon";
